// WebSystem/WebSystem/ClientApp/src/models/Formas/INV/dp03acom.ts
export interface IDp03acom {
    codigo: string;
    codigo_i?: string;
    codigo_e?: string;
    nombre?: string;
    main?: string;
    tipo?: string;
    pidecosto?: string;
    inventario?: string;
    numero?: string;
    forma?: string;
    forma2?: string;
    grupo?: string;
    iva?: string;
    items?: string;
    ocultaiva?: string;
    permitedes?: string;
    tp_credito?: string;
    tp_aplica?: string;
    integra?: boolean;
    empresa?: string;
    tpintegra?: string;
    generanc?: boolean;
    ptovta?: string;
    v_traspor?: boolean;
    c_lineas?: string;
    tp_ret?: string;
    control_it?: string;
    num_it?: number;
    tpnumera?: string;
    od?: boolean;
    ing_valdes?: boolean;
    rma?: boolean;
    iva_credi?: string;
    ctaxpagar?: boolean;
    devol?: boolean;
    actcto?: boolean;
    transp?: boolean;
    tipo2?: string;
    ajuste?: string;
    fiscal?: string;
    signo?: string;
    rt_manu?: string;
    es_produ?: string;
    at_exporta?: string;
    tiport?: string;
    celectro?: string;
    serie?: string;
    tp_op?: string;
    es_lc?: string;
    cta_provi?: string;
    tp_ent?: string;
    es_exporta?: string;
    validart?: boolean;
    activa_mod?: string;
    num_descu?: number;
    printerPDF?: string;
    es_reserva?: boolean;
    ajuSave?: boolean;
    bode_linea?: boolean;
    orden?: number;
    cSaldo?: string;
    tp_deuda?: string;
    con_ctavta?: string;
}