// WebSystem/WebSystem/ClientApp/src/models/Formas/INV/dpinvcab.ts
export interface IDpinvcab {  
    tipo: string;
    numero: string;
    numero_b: string;
    referencia?: string;
    grupo?: string;
    tipo_t?: string;
    fecha_tra?: Date;
    fecha_ven?: Date;
    bodega_d?: string;
    bodega?: string;
    prov_cli?: string;
    vendedor?: string;
    tp_precio?: string;
    total_mov?: number;
    pordes?: number;
    total_des?: number;
    poriva?: number;
    total_iva?: number;
    total_trn?: number;
    userfec?: Date;
    userfec_2?: string;
    usertim?: string;
    usercla?: string;
    comenta?: string;
    comenta1?: string;
    comenta2?: string;
    comenta3?: string;
    vta_cli?: string;
    vta_ruc?: string;
    vta_dir?: string;
    vta_fo1?: string;
    vta_fo2?: string;
    anulada?: boolean;
    fecha_anu2?: Date;
    fecha_anu?: Date;
    hora_anu?: string;
    user_anu?: string;
    abono?: number;
    fac_pro?: string;
    codigo_i?: string;
    codigo_e?: string;
    tntegracnt?: string;
    vapor?: string;
    fec_emb?: Date;
    semana?: string;
    guia_remi?: string;
    numfue?: string;
    librasne?: number;
    pesone?: number;
    pesobru?: number;
    marca?: string;
    destino?: string;
    cartonera?: string;
    cancelapto?: string;
    comprodev?: string;
    van?: number;
    cajero?: string;
    flete?: number;
    transporte?: string;
    terminos?: string;
    st?: string;
    canc_lote?: boolean;
    factura?: string;
    dirdes?: string;
    remite?: string;
    fec_ini?: Date;
    fec_fin?: Date;
    trasladox?: string;
    base_0?: number;
    base_tax?: number;
    val_trans?: number;
    porcom?: number;
    integracos?: string;
    ordcom?: string;
    cubica_t?: number;
    ano?: number;
    periodo?: number;
    catego?: number;
    programa?: number;
    cuota?: number;
    contrib?: string;
    rma?: string;
    ocompra?: string;
    s_impcadu?: string;
    s_autoriza?: string;
    od?: boolean;
    fecha_ve1?: Date;
    status_flu?: string;
    fecha_flu?: Date;
    ctacont?: string;
    total_ser?: number;
    por_comitc?: number;
    base_tc?: number;
    valcomitc?: number;
    contabi?: boolean;
    impreso?: string;
    desc_fac?: string;
    valor_rt?: number;
    pesada?: string;
    lote?: string;
    contenedor?: string;
    sellos?: string;
    export?: string;
    subpartiAn?: string;
    marcas?: string;
    guia_placa?: string;
    guia_nDestina?: string;
    guia_RucDest?: string;
    aguaje?: string;
    lrecibida?: number;
    lbasura?: number;
    rendimiento?: number;
    total_lib?: number;
    total_kil?: number;
    total_caj?: number;
    sobrante?: number;
    sobr_basura?: number;
    sobr_remite?: number;
    sobr_rendi?: number;
    peso_planta?: number;
    trans_ruc?: string;
    partida?: string;
    transporta?: string;
    llegada?: string;
    ruta?: string;
    ce_autoriz?: string;
    ce_clave?: string;
    ce_fecaut?: Date;
    ce_activo?: string;
    ce_estado?: string;
    ce_deserra?: string;
    fecha_clu?: Date;
    ap_comi?: Date;
    recibido?: string;
    fecha_rec?: Date;
    hora_rec?: string;
    user_rec?: string;
    comenta_re?: string;
    conceptox?: string;
    rs_scan?: string;
    estadoFac?: boolean;
    flete_loc?: number;
    gasto_loc?: number;
    valor_fob?: number;
    flete_int?: number;
    seguro?: number;
    valor_cif?: number;
    valor_adu?: number;
    gastos_lod?: number;
    valtransp?: number;
    valotros?: number;
    total_sub?: number;
    tipo_trans?: string;
    pais?: string;
    doc_embar?: string;
    pto_embar?: string;
    tam_conte?: string;
    num_conte?: string;
    tipo_segur?: string;
    embarque?: Date;
    liquida?: string;
    import?: string;
    come_anula?: string;
    total_ice?: number;
    csf_nombre?: string;
    csf_cedula?: string;
    csf_direcc?: string;
    cpcodigo_1?: string;
    cpValor_1?: number;
    cpcodpla_1?: string;
    cpvalpla_1?: number;
    cpcodigo_2?: string;
    cpValor_2?: number;
    cpcodpla_2?: string;
    cpvalpla_2?: number;
    cpcodigo_3?: string;
    cpValor_3?: number;
    cpcodpla_3?: string;
    cpvalpla_3?: number;
    cpcodigo_4?: string;
    cpValor_4?: number;
    cpcodpla_4?: string;
    cpvalpla_4?: number;
    porpromc?: number;
    esCotiza?: number;
    cSaldo?: string;
}