// WebSystem/WebSystem/ClientApp/src/models/Formas/INV/dp03amov.ts
export interface IDp03amov {  
    tipo: string;
    numero: string;
    linea: number;
    tipo_t: string;
    presenta?: string;
    factor?: number;
    fecha_tra: Date;
    touch?: string;
    no_parte?: string;
    consumo?: string;
    bodega?: string;
    cantidad?: number;
    despacho?: number;
    precio_u?: number;
    precio_t?: number;
    costo_u?: number;
    costo_t?: number;
    poriva?: number;
    valiva?: number;
    pordes?: number;
    valdes?: number;
    precio_ne?: number;
    cant_tou?: number;
    pvpu_tou?: number;
    pvpt_tou?: number;
    cubu_tou?: number;
    cubt_tou?: number;
    cubica_u?: number;
    cubica_t?: number;
    servicio?: string;
    van?: number;
    desitem?: number;
    tprecio?: string;
    serie?: string;
    desit?: number;
    iat?: string;
    devol?: boolean;
    cant_devol?: number;
    nservicio?: string;
    marca?: string;
    modelo?: string;
    cambia?: boolean;
    tipnc?: string;
    no_parteco?: string;
    cta_costo?: string;
    s_impcadu?: string;
    s_autoriza?: string;
    valor_fac?: number;
    factorvta?: number;
    desfacvta?: string;
    ceco?: string;
    desit2?: number;
    kgbrutos?: number;
    boxes?: number;
    no_partea?: string;
    cajetas?: number;
    talla?: string;
    clase?: string;
    calidad?: string;
    congela?: string;
    tpeso?: string;
    pesoMar?: string;
    pesoEmp?: string;
    corte?: string;
    empaque?: string;
    masterp?: string;
    total_lib?: number;
    total_kil?: number;
    total_caj?: number;
    logistica?: number;
    descabezado?: number;
    es_sobrante?: boolean;
    lote?: string;
    no_partec?: string;
    desfac?: string;
    factor2?: number;
    recet?: string;
    costFactor?: number;
    totFactor?: number;
    codFactor?: string;
    canFactor?: number;
    num_descu?: number;
    desit3?: number;
    desit4?: number;
    desit5?: number;
    desit6?: number;
    desit7?: number;
    desit8?: number;
    desit9?: number;
    desit10?: number;
    descripD1?: string;
    descripD2?: string;
    descripD3?: string;
    descripD4?: string;
    descripD5?: string;
    descripD6?: string;
    descripD7?: string;
    descripD8?: string;
    descripD9?: string;
    descripD10?: string;
    porcenta_n?: number;
    flete_loc?: number;
    gastos_loc?: number;
    valor_fob?: number;
    flete_int?: number;
    seguro?: number;
    valor_cif?: number;
    valor_aud?: number;
    gastos_lod?: number;
    valtransp?: number;
    valotros?: number;
    total_sub?: number;
    costo_uni2?: number;
    costo_list?: number;
    porICE?: number;
    partidaICE?: string;
}

export interface IDp03amovEXT extends IDp03amov {
  nombreItem?: string;  
}